event_handler/requirements.txt
amqp
python-dotenv
flask
flask-sqlalchemy
psycopg2
celery
flask_socketio